refactored version for multiple tools
