+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "dggrs"
-version = "0.4.1"
+version = "0.5.0"
 edition = "2021"
 readme = "README.md"
