+++ src/adapters/mod.rs
@@ -1,2 +1,1 @@
 pub mod dggrid;
-pub mod dggrid_service;
