+++ src/adapters/rhealpix/rhealpix.rs
@@ -1,0 +1,30 @@
+use crate::ports::dggrs::DggrsPort;
+
+pub struct RhealpixImpl;
+
+impl DggrsPort for RhealpixImpl {
+    fn whole_earth(
+        &self,
+        dggs_type: String,
+        dggs_res_spec: u8,
+        densify: bool,
+        bbox: Option<Vec<Vec<f64>>>,
+    ) -> CellsGEO {
+        // call rHEALPix here and return CellsGEO
+    }
+    fn from_point(
+        &self,
+        dggs_type: String,
+        dggs_res_spec: u8,
+        point: Point,
+        densify: bool,
+    ) -> CellsGEO {
+        // etc.
+    }
+    fn coarse_cells(&self, ...) -> CellsGEO {
+        // ...
+    }
+    fn single_zone(&self, ...) -> CellsGEO {
+        // ...
+    }
+}
