+++ src/dggrs/mod.rs
@@ -1,0 +1,1 @@
+pub use crate::factory::dggrs_factory::get;
