+++ src/factory/mod.rs
@@ -1,0 +1,1 @@
+pub mod dggrs_factory;
